Try to fix hippo_dimension_reduction error when SCE count is a sparse matrix
The original error came out from hippo_dimension_reduction is when the count matrix in SCE object being
a sparse matrix such as 'dgeMatrix'. I added a few lines of code to convert the matrix back to the base
matrix type in R so that umap::umap() / Rtsne::Rtsne() can process it.

I haven't checked if there is anything else I should consider for making this change.
Hope this commit won't break anthing.