Try to fix hippo_dimension_reduction error when SCE count is a sparse matrix
The original error came out from hippo_dimension_reduction is when the count matrix in SCE object being
a sparse matrix such as 'dgeMatrix'. I added a few lines of code to convert the matrix back to the base
matrix type in R so that umap::umap() / Rtsne::Rtsne() can process it.

I haven't check whether this change is going to break anything else and hope it won't.